ecommerce-site/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
stripe==5.5.0
alembic==1.12.1
Werkzeug==2.3.7
